# api/openapi.yaml
openapi: 3.0.3
info:
  title: Our Sell API
  version: 1.0.0
  description: |
    Secure, versioned authentication and session-management API. Refresh
    tokens are opaque, rotated, and stored only as SHA-256 hashes on the server.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/wallet:
    get:
      operationId: getWallet
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current authenticated user wallet summary.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/Wallet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/wallet/ledger:
    get:
      operationId: listWalletLedger
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Immutable wallet movement history.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/WalletLedgerResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/wallet/topups:
    post:
      operationId: createWalletTopup
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WalletTopupInput' }
      responses:
        '201':
          description: Dynamic KHQR created for the authenticated customer.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/WalletTopup' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/wallet/topups/{topupID}:
    get:
      operationId: getWalletTopup
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: topupID, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Top-up status after server-side Bakong verification.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/WalletTopup' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/payments/bakong/webhook:
    post:
      operationId: verifyBakongTopupWebhook
      security: []
      description: Optional signed internal relay. The API still verifies the payment against Bakong before crediting the wallet.
      parameters:
        - name: X-Bakong-Webhook-Secret
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reference: { type: string }
                qr_md5: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '400': { $ref: '#/components/responses/ValidationError' }
  /api/v1/admin/wallet/topups:
    get:
      operationId: adminListWalletTopups
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string, maxLength: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [pending_payment, confirmed, failed, expired, reversed] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Wallet top-up operations view. Requires the admin role.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/AdminWalletTopupResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Message:
      description: Operation result.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Message' }
    ValidationError:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request
              fields: { email: Invalid email address }
...
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    DependencyUnavailable:
      description: Required dependency is unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Wallet:
      type: object
      required: [id, user_id, currency, available_amount, held_amount, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        currency: { type: string, enum: [USD] }
        available_amount: { type: number, format: double, minimum: 0, example: 25.00 }
        held_amount: { type: number, format: double, minimum: 0, example: 6.00 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WalletTopupInput:
      type: object
      required: [amount_usd]
      additionalProperties: false
      properties:
        amount_usd: { type: number, format: double, minimum: 1, example: 10.00 }
    WalletTopup:
      type: object
      required: [id, wallet_id, user_id, reference, amount_usd, currency, qr_payload, qr_md5, status, expires_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        wallet_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        reference: { type: string }
        amount_usd: { type: number, format: double, minimum: 0, example: 10.00 }
        currency: { type: string, enum: [USD] }
        qr_payload: { type: string, description: Dynamic KHQR payload to render as a QR code. }
        qr_md5: { type: string, description: QR MD5 used for server-side transaction verification. }
        provider_transaction_id: { type: string }
        status: { type: string, enum: [pending_payment, confirmed, failed, expired, reversed] }
        failure_reason: { type: string }
        expires_at: { type: string, format: date-time }
        paid_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WalletLedgerEntry:
      type: object
      required: [id, wallet_id, user_id, entry_type, reference, available_delta_usd, held_delta_usd, available_balance_usd, held_balance_usd, description, created_at]
      properties:
        id: { type: string, format: uuid }
        wallet_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        entry_type: { type: string, enum: [topup_confirmed, topup_reversed, order_hold, order_release, order_spent, admin_adjustment] }
        reference: { type: string }
        available_delta_usd: { type: number, format: double }
        held_delta_usd: { type: number, format: double }
        available_balance_usd: { type: number, format: double, minimum: 0 }
        held_balance_usd: { type: number, format: double, minimum: 0 }
        description: { type: string }
        created_at: { type: string, format: date-time }
    WalletLedgerResult:
      type: object
      required: [entries, page, page_size, has_more]
      properties:
        entries:
          type: array
          items: { $ref: '#/components/schemas/WalletLedgerEntry' }
        page: { type: integer }
        page_size: { type: integer }
        has_more: { type: boolean }
    AdminWalletTopup:
      allOf:
        - $ref: '#/components/schemas/WalletTopup'
        - type: object
          required: [user_name, user_email]
          properties:
            user_name: { type: string }
            user_email: { type: string, format: email }
    AdminWalletTopupResult:
      type: object
      required: [topups, total, page, page_size, total_pages]
      properties:
        topups:
          type: array
          items: { $ref: '#/components/schemas/AdminWalletTopup' }
        total: { type: integer, format: int64 }
        page: { type: integer }
        page_size: { type: integer }
        total_pages: { type: integer }
...
    Message:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }